requirements.txt
Flask
PyMuPDF
python-docx
google-generativeai
python-magic
pdfplumber